fix: rename deps.ts
