+++ puppeteer/src/components/createZoomMeetingUsingChromeExtension.ts
@@ -1,4 +1,4 @@
-import { Page } from "../mod.ts";
+import { Page } from "../deps.ts";
 import waitFor from "../utils/waitFor.ts";
 
 const SAVE_BUTTON = "div.BTotkb.JaKw1 > div:nth-child(4) > div";
