+++ puppeteer/src/components/createZoomMeetingUsingZoomAddon.ts
@@ -1,6 +1,6 @@
 import clickGoogleCalendarAddonSaveButton from "../components/clickGoogleCalendarAddonSaveButton.ts";
 import openAccordion from "../components/openAccordion.ts";
-import { Page } from "../mod.ts";
+import { Page } from "../deps.ts";
 import waitFor from "../utils/waitFor.ts";
 
 const ZOOM_BUTTON =
