+++ puppeteer/src/createMeetings.ts
@@ -1,4 +1,4 @@
-import { Page } from "./mod.ts";
+import { Page } from "./deps.ts";
 import { testCases } from "./settings.ts";
 import waitFor from "./utils/waitFor.ts";
